feat: Add ResponsiveLayout widget
This commit adds a new widget, ResponsiveLayout, that displays different content depending on the screen size. The ResponsiveLayout widget takes three required arguments:

    phone: The widget that should be displayed on phones.
    tablet: The widget that should be displayed on tablets.
    desktop: The widget that should be displayed on desktops.

The ResponsiveLayout widget uses the LayoutBuilder widget to determine which widget to display based on the screen size. The LayoutBuilder widget takes a builder function that is called with two arguments: the BuildContext and the BoxConstraints of the LayoutBuilder. The builder function should return the widget that should be displayed for the given constraints.

In the ResponsiveLayout widget, the builder function checks the maxWidth property of the BoxConstraints object. If the maxWidth is less than or equal to 480, the phone widget is returned. If the maxWidth is greater than 480 and less than or equal to 768, the tablet widget is returned. Otherwise, the desktop widget is returned.

The ResponsiveLayout widget can be used to create a more responsive user experience by displaying different content on different screen sizes.