feat: add light_logo.svg and dark_logo.svg
This commit adds two SVG files, light_logo.svg and dark_logo.svg, to the project. These files will be used as the logo assets for the light and dark themes respectively. The logos are designed to complement the different themes and enhance the overall visual presentation of the app.