feat: add withHover extension for enabling hover functionality on a widget
This commit introduces an extension named HoverExtension on the Widget class in the Flutter codebase. The extension provides a getter withHover that allows any widget to be wrapped with hover functionality. When a widget is wrapped using withHover, a MouseRegion widget is added around it, changing the mouse cursor to SystemMouseCursors.click when the widget is hovered over.

The withHover extension enhances reusability and simplifies the process of enabling hover effects in Flutter applications. By calling .withHover on a widget, developers can easily enable the hover behavior without manually writing the MouseRegion widget code.

import 'package:flutter/material.dart';

class MyButton extends StatelessWidget {
  @override
  Widget build(BuildContext context) {
    return Container(
      // Widget code...
    ).withHover;
  }
}

This commit improves the developer experience and promotes consistency in hover effects throughout the application.